FIX: Update hooks attributes that is not correct type guard
